cleaned and added a sucess view
